Moved all files from Wox to Powerlauncher
