+++ src/modules/launcher/PowerLauncher/Helper/ErrorReporting.cs
