+++ src/modules/launcher/PowerLauncher/Helper/KeyboardHelper.cs
