+++ src/modules/launcher/PowerLauncher/Helper/SingleInstance.cs
