+++ src/modules/launcher/PowerLauncher/Helper/WindowsInteropHelper.cs
