+++ src/modules/launcher/PowerLauncher/PublicAPIInstance.cs
