+++ src/modules/launcher/PowerLauncher/ReportWindow.xaml
