+++ src/modules/launcher/PowerLauncher/Settings.cs
