+++ src/modules/launcher/PowerLauncher/Storage/HistoryItem.cs
