+++ src/modules/launcher/PowerLauncher/Storage/QueryHistory.cs
