+++ src/modules/launcher/PowerLauncher/Storage/TopMostRecord.cs
