+++ src/modules/launcher/PowerLauncher/Storage/UserSelectedRecord.cs
