+++ src/modules/launcher/PowerLauncher/ViewModel/MainViewModel.cs
