+++ src/modules/launcher/PowerLauncher/ViewModel/RelayCommand.cs
