+++ src/modules/launcher/PowerLauncher/ViewModel/ResultsViewModel.cs
