+++ src/modules/launcher/PowerLauncher/ViewModel/SettingWindowViewModel.cs
